+++ add.js
@@ -1,6 +1,8 @@
 
+var auto = require("./auto");
+
 function add () {
     return Array.prototype.reduce.call(arguments, function (a, b) { return a + b; });
 }
 
-module.exports = add;
+module.exports = auto(add, 2);