+++ divide.js
@@ -1,6 +1,8 @@
 
+var auto = require("./auto");
+
 function divide () {
     return Array.prototype.reduce.call(arguments, function (a, b) { return a / b; });
 }
 
-module.exports = divide;
+module.exports = auto(divide, 2);