+++ mod.js
@@ -1,6 +1,8 @@
 
+var auto = require("./auto");
+
 function mod () {
     return Array.prototype.reduce.call(arguments, function (a, b) { return a % b; });
 }
 
-module.exports = mod;
+module.exports = auto(mod, 2);