+++ multiply.js
@@ -1,6 +1,8 @@
 
+var auto = require("./auto");
+
 function multiply () {
     return Array.prototype.reduce.call(arguments, function (a, b) { return a * b; });
 }
 
-module.exports = multiply;
+module.exports = auto(multiply, 2);