+++ subtract.js
@@ -1,6 +1,8 @@
 
+var auto = require("./auto");
+
 function subtract () {
     return Array.prototype.reduce.call(arguments, function (a, b) { return a - b; });
 }
 
-module.exports = subtract;
+module.exports = auto(subtract, 2);