+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog for enjoy-core
 
+## 1.2.0 (2016-09-05)
+
+ * Make `auto` functions work with more than one argument
+ * add optional arity parameter to `auto` to make it work with functions that rely on the `arguments` object
+ * Make functions with more than one argument `auto`.
+
 ## 1.1.0
 
  * Add a `fallback(fn, instead)` function, which is a functional try-catch